requirements.txt
numpy>=1.26
pandas>=2.1
matplotlib>=3.8
scikit-learn>=1.4
Pillow>=10.0
opencv-python>=4.9
tqdm>=4.66
torch>=2.2
torchvision>=0.17
umap-learn>=0.5
